requirements/base.txt
Django==4.2.1
django-zipfile==0.4.0
langchain==0.0.295
unstructured==0.6.2
pdfminer.six==20221105
qdrant-client==1.1.6
openai==0.27.5
tiktoken==0.3.3
django-storages==1.13.2
boto3==1.26.123
django-tailwind==3.5.0
celery==5.2.7
redis==4.5.4
kombu==5.2.4
django-celery-results==2.5.1
paradigm-client @ git+https://github.com/lightonai/paradigm-client.git@fa32216b1080131fe3776b7ba75124f6ebe6db5a
accelerate==0.19.0
bitsandbytes==0.38.1
transformers==4.29.1
sentencepiece==0.1.99

# see https://github.com/Muennighoff/sgpt/issues/14#issuecomment-1529010350
sentence-transformers @ git+https://github.com/Muennighoff/sentence-transformers.git@5c025b34ea665a98cbac028e45d1cea91dc33b86

channels==4.0.0
channels-redis==4.1.0

djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.2

# for vertex ai
google-cloud-aiplatform==1.25.0

# for anonymization
spacy==3.6.1
https://github.com/explosion/spacy-models/releases/download/fr_core_news_lg-3.6.0/fr_core_news_lg-3.6.0.tar.gz
pybloom-live==4.0.0

# for pdf text extraction
pdfplumber==0.9.0
nltk==3.8.1